MOB-6442 -> Update the d-types for max space and usage summary, to fix JSON PE -> Bump version

+++ components/src/test/java/com/constantcontact/v2/library/LibraryInfoTest.java
@@ -25,7 +25,7 @@ public class LibraryInfoTest {
 
     private static final int MAX_FREE_FILE = 5;
 
-    private static final int MAX_PREMIUM_SPACE = 100;
+    private static final long MAX_PREMIUM_SPACE = 100;
 
     private static final int MAX_UPLOAD_SIZE = 1;
